Do not use param struct everywhere
In the future we should move to a non-static version of this lib
that will allow us to set run time values for the simulation
parameters. This is of the utmost importance in order to run
optimization algorithms and/or other programs that require trying
out different parameters at each run.